requirements.txt
openai>=1.50.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.70.0
pyyaml>=6.0
python-dotenv>=1.0
feedparser>=6.0
httpx>=0.27.0
aiohttp>=3.10.0
Pillow>=10.0
jinja2>=3.1
# Newsletter API
fastapi>=0.110.0
uvicorn>=0.27.0
# TikTok 動画生成 (ffmpeg が別途必要: brew install ffmpeg / apt install ffmpeg)
moviepy>=2.0
numpy>=1.24
